services/reranker-service/pyproject.toml
[tool.poetry]
name = "reranker-service"
version = "1.0.0"
description = "RAG reranker service — GPU CrossEncoder"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = { extras = ["standard"], version = "^0.29.0" }
pydantic-settings = "^2.0"
sentence-transformers = "^3.0"
rag-common = { path = "../../packages/rag-common", develop = true }

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-asyncio = "^0.23"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
